add Wikipedia inspired bibliography style (#29)
* Create README.md

* Add the bibliography file

Add the scheme file for the bibliography style

* Upload OpenAccess Icons

The Icons were created by the author and are in the public domain. They are created to have a dpi of approximately 300.

* Update README.md

Add description to README

* Add screenshot for README

* Update README.md

Fix some minor typos

* Update README.md

Oops, I accidentally forgot to credit the original source of the Icons